requirements.txt
asgiref==3.7.2
distlib==0.3.7
Django==4.2.3
djangorestframework==3.14.0
filelock==3.13.1
platformdirs==3.11.0
pytz==2023.3.post1
sqlparse==0.4.4
tzdata==2023.3
virtualenv==20.24.6
